+++ hooks/field/string/useFieldDateChangeEventCallback.ts
@@ -1,31 +1,24 @@
 // Copyright (c) 2022. Heusala Group Oy <info@heusalagroup.fi>. All rights reserved.
 
-import { ChangeEvent, Dispatch, SetStateAction, useCallback } from "react";
+import { Dispatch, SetStateAction, useCallback } from "react";
+import { moment } from "../../../../core/modules/moment";
 import { FieldChangeCallback, useFieldChangeCallback } from "../useFieldChangeCallback";
 
-/**
- *
- * @param identifier
- * @param setValue
- * @param change
- */
-export function useFieldStringChangeEventCallback (
+export function useFieldDateChangeEventCallback (
     identifier: string,
     setValue: Dispatch<SetStateAction<string>>,
-    change: FieldChangeCallback<string | undefined> | undefined
+    change: FieldChangeCallback<string | undefined> | undefined,
+    dateFormat?: string
 ) {
     const changeCallback = useFieldChangeCallback<string>(identifier, change);
     return useCallback(
-        (event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-            if ( event ) {
-                event.preventDefault();
-                event.stopPropagation();
-            }
-            const eventTargetValue = event?.target?.value ?? '';
+        (value: string) => {
+            const eventTargetValue = moment(value).format(dateFormat) ?? '';
             setValue(eventTargetValue);
             changeCallback(eventTargetValue);
         },
         [
+            dateFormat,
             changeCallback,
             setValue
         ]
